Reorganize the InsightWidgetFiles
JIRA: RAIL-4391
