+++ libs/sdk-ui-dashboard/src/presentation/widget/widget/InsightWidget/index.ts
@@ -1,0 +1,3 @@
+// (C) 2022 GoodData Corporation
+
+export * from "./RenderModeAwareDashboardInsightWidget";
